Added FetchHandler Component
Added a Component to handle fetch request so that we can start adding other requests easily(planets, films, etc)
