+++ src/LoadingScreen.js
@@ -1,6 +1,8 @@
 import React from 'react'
 
 const LoadingScreen = (props) => (
+  props.error ?
+  <h1>ERROR GETTING DATA</h1> :
   <div className="loading-container">
     <h2>Fetching data</h2>
     <div className="lds-ellipsis">
